exp: pretraining hr gram loss with learned downsampling of hr repr.
Instead of interpolation, use linear projection to downsample.
What is projected to what may need to be looked at. What should
happen is all the patches repr. from the HR image corresponding to one
patch in the LR image should be projected to be one patch, which is
used to compute loss with the corresponding LR patch repr.